requirements.txt
qq# Core OMR Evaluation
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
Pillow==10.0.0
PyMuPDF==1.23.5
pdfplumber==0.9.0

# Web Application
fastapi==0.103.1
uvicorn[standard]==0.23.2
streamlit==1.25.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Data Processing
pandas==2.0.3
openpyxl==3.1.2

# Image Processing
imutils==0.5.4
matplotlib==3.7.2

# Utilities
python-dotenv==1.0.0
pydantic==2.1.1
joblib==1.3.2

# Additional dependencies
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
plotly==5.17.0
requests==2.31.0